backend/requirements.txt
# Core dependencies for all Lambda functions
boto3>=1.28.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
requests>=2.31.0
